zentratech/main/serializers.py
# serializers.py
from rest_framework import serializers
from .models import Interest, ChatMessage

class InterestSerializer(serializers.ModelSerializer):
    class Meta:
        model = Interest
        fields = '__all__'

class ChatMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ChatMessage
        fields = '__all__'
