requirements.txt
Django==4.1.5
djangorestframework==3.13.1
django-cors-headers==3.13.0